+++ jsr.json
@@ -19,6 +19,7 @@
     "./ip-restriction": "./src/middleware/ip-restriction/index.ts",
     "./cache": "./src/middleware/cache/index.ts",
     "./route": "./src/helper/route/index.ts",
+    "./mount": "./src/middleware/mount/index.ts",
     "./cookie": "./src/helper/cookie/index.ts",
     "./accepts": "./src/helper/accepts/index.ts",
     "./compress": "./src/middleware/compress/index.ts",
